TST: restructure tests and skip VDS tests on old hdf5
