+++ .travis.yml
@@ -21,38 +21,38 @@ env:
   global:
     - HDF5_CACHE_DIR=$HOME/.cache/hdf5
   matrix:
-      #    - TOXENV=py26-test-deps
-      #    - TOXENV=py27-test-deps
-      #    - TOXENV=py33-test-deps
-      #    - TOXENV=py34-test-deps
-      #    - TOXENV=py26-test-mindeps
-      #    - TOXENV=py27-test-mindeps
-      #    - TOXENV=py33-test-mindeps
-      #    - TOXENV=py34-test-mindeps
-      ##    - TOXENV=py35-test-deps
-      ##    - TOXENV=py35-test-mindeps
-      ##    - TOXENV=py36-test-deps
-      ##    - TOXENV=py36-test-mindeps
-      ##   commented out because of https://github.com/travis-ci/travis-ci/issues/4794
-      #    - TOXENV=py26-test-deps
-      #      TOX_TESTENV_PASSENV=LANG LC_ALL
-      #      LANG=C
-      #      LC_ALL=C
-      #    - TOXENV=py27-test-deps
-      #      TOX_TESTENV_PASSENV=LANG LC_ALL
-      #      LANG=C
-      #      LC_ALL=C
-      #    - TOXENV=py33-test-deps
-      #      TOX_TESTENV_PASSENV=LANG LC_ALL
-      #      LANG=C
-      #      LC_ALL=C
-      #    - TOXENV=py34-test-deps
-      #      TOX_TESTENV_PASSENV=LANG LC_ALL
-      #      LANG=C
-      #      LC_ALL=C
-      #    - TOXENV=py27-test-deps
-      #      HDF5_VERSION=1.10.1
-      #      HDF5_DIR=$HDF5_CACHE_DIR/$HDF5_VERSION
+    - TOXENV=py26-test-deps
+    - TOXENV=py27-test-deps
+    - TOXENV=py33-test-deps
+    - TOXENV=py34-test-deps
+    - TOXENV=py26-test-mindeps
+    - TOXENV=py27-test-mindeps
+    - TOXENV=py33-test-mindeps
+    - TOXENV=py34-test-mindeps
+#    - TOXENV=py35-test-deps
+#    - TOXENV=py35-test-mindeps
+#    - TOXENV=py36-test-deps
+#    - TOXENV=py36-test-mindeps
+#   commented out because of https://github.com/travis-ci/travis-ci/issues/4794
+    - TOXENV=py26-test-deps
+      TOX_TESTENV_PASSENV=LANG LC_ALL
+      LANG=C
+      LC_ALL=C
+    - TOXENV=py27-test-deps
+      TOX_TESTENV_PASSENV=LANG LC_ALL
+      LANG=C
+      LC_ALL=C
+    - TOXENV=py33-test-deps
+      TOX_TESTENV_PASSENV=LANG LC_ALL
+      LANG=C
+      LC_ALL=C
+    - TOXENV=py34-test-deps
+      TOX_TESTENV_PASSENV=LANG LC_ALL
+      LANG=C
+      LC_ALL=C
+    - TOXENV=py27-test-deps
+      HDF5_VERSION=1.10.1
+      HDF5_DIR=$HDF5_CACHE_DIR/$HDF5_VERSION
     - TOXENV=docs
 
 matrix:
