+++ h5py/tests/hl/test_vds/__init__.py
@@ -1,9 +1,6 @@
 from __future__ import absolute_import
 
-from .test_eiger_high_level import *
-from .test_eiger_low_level import *
-from .test_excalibur_high_level import *
-from .test_excalibur_low_level import *
-from .test_percival_high_level import *
-from .test_percival_low_level import *
+
 from .test_virtual_source import *
+from .test_highlevel_vds import *
+from .test_lowlevel_vds import *
